pyproject.toml
[tool.poetry]
name = "imagebase"
version = "0.1.0"
description = "image processing"
authors = ["Hoo"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
pillow = "^8.1.0"
aiofiles = "^0.6.0"
ujson = "^4.0.2"
python-multipart = "^0.0.5"
loguru = "^0.5.3"
tortoise-orm = "^0.16.21"
aiosqlite = "^0.16.1"
asyncpg = "^0.22.0"
aiomysql = "^0.0.21"
aioredis = "^1.3.1"
numpy = "^1.20.1"
scipy = "^1.6.1"
pywavelets = "^1.1.1"

[tool.poetry.dev-dependencies]
# pytest
requests = "^2.25.1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-asyncio = "^0.14.0"
httpx = "^0.16.1"
asynctest = "^0.13.0"
flake8 = "^3.8.4"
isort = "^5.7.0"
autopep8 = "^1.5.5"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true  # 设置清华镜像为默认回调, 以避免 poetry 在某些缺省操作里又回调到 pypi

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_third_party = ["fastapi", "pydantic", "starlette"]